whiteboard renders properly using a react component
